refactor: relocate editor tooling under cartographer
